CB-12407: Increase FreeIPA certificate serial number max value
The FreeIPA certificate serial number's is an ASN.1 integer. ANS.1
intergers are unbounded. The java code was increased from an int to a
long. While this doesn't fix the problem in all cases, in practise it
should be sufficient. When FreeIPA is installed, it will pick a range
and use consecutive serial number for that instance of FreeIPA.

This was tested using a local deployment of cloudbreak that had an out
of range number in the serial number.

+++ freeipa-client/src/main/java/com/sequenceiq/freeipa/client/model/Cert.java
@@ -10,7 +10,7 @@ public class Cert {
     private String cacn;
 
     @JsonProperty("serial_number")
-    private int serialNumber;
+    private long serialNumber;
 
     @JsonProperty("valid_not_before")
     private String validNotBefore;
@@ -40,11 +40,11 @@ public class Cert {
         this.cacn = cacn;
     }
 
-    public int getSerialNumber() {
+    public long getSerialNumber() {
         return serialNumber;
     }
 
-    public void setSerialNumber(int serialNumber) {
+    public void setSerialNumber(long serialNumber) {
         this.serialNumber = serialNumber;
     }
